Complete validation and exception handling

+++ src/main/java/com/ems/dto/EmployeeRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class EmployeeRequest {
     @NotBlank(message = "First name is required")
@@ -26,5 +28,5 @@ public class EmployeeRequest {
     private String designation;
 
     @Positive(message = "Salary must be greater than zero")
-    private Double salary;
+    private BigDecimal salary;
 }
