+++ src/main/java/com/ems/entity/Employee.java
@@ -3,6 +3,8 @@ package com.ems.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "employees")
 @Getter
@@ -30,5 +32,5 @@ public class Employee {
 
     private String designation;
 
-    private double salary;
+    private BigDecimal salary;
 }
